Another daemonization attempt
The previous fix to get the process to stop crashing didn't appear to
work. The reason for this is that the nohup utility will ignore a hangup
signal, so logging out won't have any impact. However, the reason the
program crashes isn't because of that, but because the SOCKET that it
listens to hangs up.

Forever is a npm program installed globally that watches processes and
restarts them as soon as they die, fixing the problem permanently.